Add support for logging in with resume token.
If the token is not valid then it will automatically retry using
the username/email and password. this way there is no need to
check if the token is expired, although you end up making an extra
call.

Tokens are automatically saved for use when a reconnection is
required, and you can also pass an optional token into the login
method so that your app can persist it between runs.